added grep check of current file to prevent duplicates

+++ tools/sudoadd.sh
@@ -17,7 +17,9 @@ sudo rm -f "$sudoers_testfile"
 test $rc -ne 0 && exit $?
 
 # Add it for real
-echo "$sudoers_data" | sudo tee -a "$sudoers_file" &>/dev/null
+if [ ! grep -s "$sudoers_data" "$sudoers_file" &>/dev/null ]; then
+  echo "$sudoers_data" | sudo tee -a "$sudoers_file" &>/dev/null
+fi
 
 # Lastly execute command
 sudo $*
